+++ src/damionjunk/nlp.clj
@@ -1,9 +1,28 @@
 (ns damionjunk.nlp
-  (:require [damionjunk.nlp.stanford :as nlp]))
+  (:require [damionjunk.nlp.stanford :as nlp]
+            [damionjunk.nlp.cmu-ark  :as ark]))
 
 ;; Nothing here but some examples for now.
 
 (comment
+
+  ;; This example uses CMU's ark-tweet-nlp to give us a more accurate POS tagging:
+  (ark/tag "ikr? u r my best friend. :) LOL amirite? #funzone")
+
+  ;; =>
+  ;; ({:token "ikr", :pos "!"}
+  ;;  {:token "?", :pos ","}
+  ;;  {:token "u", :pos "O"}
+  ;;  {:token "r", :pos "V"}
+  ;;  {:token "my", :pos "D"}
+  ;;  {:token "best", :pos "A"}
+  ;;  {:token "friend", :pos "N"}
+  ;;  {:token ".", :pos ","}
+  ;;  {:token ":)", :pos "E"}
+  ;;  {:token "LOL", :pos "!"}
+  ;;  {:token "amirite", :pos "!"}
+  ;;  {:token "?", :pos ","}
+  ;;  {:token "#funzone", :pos "#"})
 
 
   (nlp/sentiment-maps "I'm not sure if I like the movie. I'm quite certain I do not like the movie. I really loved the movie, it was great.")