+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 0.3.0
+
+* Added POS tagging from CMU's ark-tweet-nlp for more accurate POS on intarwebspeek.
+* Added some basic tests using Expectations
+
 # 0.2.0
 
 * Added a function for full annotation (ner/pos/sentiment)